Batch 10: reproduce reference View Gem presentation (R-17, Y-14)
Replace the bit-test tile coloring and HUD-clipped overworld/dungeon gem
rendering with a portable, host-tested semantic layer ported from the
reference buildGemView/GEM_CATEGORY/initChunkOrigin, and paint it through
the live EGA palette instead of invented colors.

- native/core/{include,src}/gem_view.{h,cpp}, gem_category.inc: new
  ESP-free semantic layer (terrain category classification, chunk-origin-
  anchored overworld window vs. fixed full town window, dungeon 22x22
  flood-fill with reference-correct wall/door/secret-door blocker rules
  that never consult discovery state, unlike movement passability).
- gem_view_test.cpp: RED->GREEN host coverage for classification, a mixed
  map fixture, N/S/E/W orientation, marker placement, full-square output,
  and blocker/wrap bounds behavior.
- generate-gem-category-table.ts: drift-guards the ported category table
  against gemmap-overworld.ts.
- native_renderer.cpp/.h: render_world_gem_view/render_dungeon_gem_view
  now only paint the semantic layer's output; no reference-parity logic
  left in device drawing code.
- tdeck_board.{h,cpp}, alpha_runtime.cpp: show_alpha gained a
  full_square_viewport mode so the gem view's full 176x176 square is
  never cropped by, or overdrawn with, the sky/wind HUD strips.

Also fixes two pre-existing build-hygiene issues (a sign-conversion cast
in dungeon_art.cpp, a string-concatenation warning in debug_labels.cpp)
required to get a clean host build in this environment's current
toolchain; both are behavior-preserving. A newly-surfaced but pre-existing
frontend struct-padding memcmp sensitivity (reproduced identically on the
untouched pre-batch tree) and the known R-21 gameplay_parity mismatch 2034
remain open and unrelated.

Host suite: 77 total, 75 pass, 2 known pre-existing failures. Firmware
(idf.py build) PASS, 852,916 B image, 19% app-partition free. Hardware
visual verification is still required and tracked in the audit.

+++ native/core/src/dungeon_art.cpp
@@ -287,8 +287,8 @@ size_t dungeon_art_blits(const DungeonDrawOp &op, const DungeonArtCatalog &catal
         if (op.bank >= kDungeonMonBanks || capacity < 2) return 0;
         const int di = int(op.depth) - 1;
         const int16_t y = kMonsterY[op.ceiling ? 1 : 0][di];
-        const uint8_t left_image = uint8_t((phase & 1u) * 3 + di);
-        const uint8_t right_image = uint8_t(((phase >> 1) & 1u) * 3 + di);
+        const uint8_t left_image = uint8_t((phase & 1u) * 3 + unsigned(di));
+        const uint8_t right_image = uint8_t(((phase >> 1) & 1u) * 3 + unsigned(di));
         const DungeonArtImage &left_dims = catalog.mon[left_image];
         const DungeonArtImage &right_dims = catalog.mon[right_image];
         if (left_dims.w == 0 || right_dims.w == 0) return 0;
